requirements.txt
asgiref==3.7.2
bcrypt==4.1.2
certifi==2024.6.2
charset-normalizer==3.3.2
Django==5.0
django-filter==24.2
djangorestframework==3.15.1
idna==3.7
mysql-connector==2.2.9
mysql-connector-python==8.2.0
nats-py==2.7.2
paho-mqtt==1.6.1
protobuf==4.21.12
psutil==5.9.7
requests==2.32.3
sqlparse==0.4.4
urllib3==2.2.1
mysql
